Removing bidding logic from logic.cpp, making it's own file. Added unit test.
While adding logic to merge cards for the winning bidder, I decided to
slit out the bidding logic from the Logic class, to simplify it. As part
of that, I added the Score class, which manages the bidding and team
scores.

Added unit testing for this new class.

The build is broken with this checkin. The changes to the bidding logic
is in progress.