Save and restore bidding dialog in the position where the player moved it
This saves the bidding dialog box position as a convenience to the user.
This will remember the position for a round of bidding, but on the next
hand, it'll start over at 0,0.

In the future, when we save game settings to the file system, we can
keep the position across bidding hands, and even games.