Start of merge cards logic, and changing card flip vs selection
While adding merge cards logic, it dawned on me the card flipping and
selection should be handled by MainWindow. Other classes shouldn't need
to manage the desktop for these generic actions.

To start, created an array that saves a pointer to a card for each
location, so we don't need to constantly pass a card to do something
minor like flip the card, or select it.

Some updates to the GUI to make the kitty cards the correct size, which
was a problem when selecting them.