Made flipping, selecting of cards local to MainWindow
Maybe this is moving too much game logic into the GUI class, but this
seems a rather generic thing to do, and using signals/slots and letting
logic handle it was messy.

Added m_playerId to Player, so it's easy for player functions to know
their Id when updating the GUI.

Removed Logic::AddCardToPlayer. This overhead was uncessary, and should
be in the Player class. Moved it to Player::AddCardToHand.