Merging of bidder and kitty complete. Moved selection, flipping of cards to MainWindow.
Finished merging of cards between player and kitty. Added a custom
dialog to show # of selected cards.
Some restructuring of the code. MainWindow now creates the players and
the card deck, and manages selecting and flipping of cards, not Logic.
These are generic card actions regardless of the game.
Moved card rotation parameter to Player, not card, as it's not specific
to cards, at least not for this game.
Removed some build warnings. Renamed game_settings to show it's specific
to this game.