Start of adding card play for hands
Started adding hand play capability. Currently, each player can choose a
card to play, which is checked for validity. Scoring of a trick isn't
done, so it doesn't determine who won the trick and move onto the next
player.
Other changes include:
1. Updated GUI to put a transparent card at an empty place, so we can
show an empty table.
2. Moved selection and orientation properties of the card to Player,
from Card. Player is the class to use to control the look on the GUI.
3. Moved the card back images to Deck, rather than duplicating this in
each card.
4. Updated when cards are face-up on the table, so the user only sees
the cards they'd see if they were that player.
5. Organized functions in Player to make them more logically ordered.